Optimize collision avoidance by storing nearby tanks.
In collision_avoidance and can_reverse, compute and store an array of
nearby vehicles and only check against them.  In my test map with 10
tanks, that reduces the time in collision avoidance from about 0.8ms per
tick to 0.4ms.  In scenarios with larger numbers of tanks it should help
more because collision avoidance is quadratic.