Refactor: Move all per-tick logic into a single iteration.
First, iteration is fairly expensive, so consolidating is itself a
performance win (although I need to put back some of the tick
throttling).  Second, having almost all the logic in one function will
make it easier to do certain kinds of throttling based on the overall
state of the force.