Fix inconsistency: players were added by name but died by type.
The code that added player controllers looked at the 'name' attribute,
but the code that watched entities die looked at 'type'.  I think there
is (in the base game) only one sort of entity with name or type
"player", but the mod should consistently use one or the other.  It now
uses 'name', as the comments on the controller structure indicate.