Update to be compatible with Factorio 0.16.
The 'passenger' attribute of entities was replaced by 'get_driver()'.

Now, 'icon_size' is a required attribute when 'icon' is present.
https://forums.factorio.com/46437

The tank now has a high-res version of its graphics, so the blue tint
for Robotanks must be applied to both the normal and high-res versions.

It is now possible to hide the ammo type indicator that turrets
normally show when detailed-info is on (Alt key by default), so that is
now done for robotank-turret-entity.

The tank had its no-fuel alert icon shifted, so I applied the same shift
to the hidden turret so that its no-ammo icon will be in the same place.

It is now possible to suppress the alert when a turret is attacking, so
that is done for robotank-turret-entity.

RoboTank technology now depends on advanced-electronics-2.  This is not
really related to the upgrade to 0.16, just an opportunistic correction
of a previous oversight, as the recipe requires blue circuits so those
should already have been researched.

It seems that the 'order' attribute of the robotank technology now works
as I intend, so I deleted the comment saying otherwise.