Fix bug: Mod data in save file was huge due to nearby_controllers.
Although I never saw this in 0.15, in 0.16 when I saved a map with 40
tanks and then loaded it, Factorio gave the error "function or
expression too complex".  I tracked this down to the fact that the
nearby_controllers array is in 'global', and was using more than half
a megabyte with 40 tanks on the map.  The fix is to make that data not
be persistent; it can be recalculated easily after loading.

This change unfortunately means that the data for a "controller" is
now split between the main controller object and the auxiliary
'unit_number_to_nearby_controllers' array.