todo.txt: Remove optimization idea for moving turrets less often.
While it is true that eliminating the ammo icon makes the jiggle
less noticeable, it is still noticeable when tanks are firing, as
the muzzle flash graphic jiggles.  I can nearly eliminate that by
updating once per 2 ticks, but then the performance gain is
negligible.  So I just won't do it.

Also remove thing about alerts since I did that already.

And the research change for the same reason.