Fix bug: Crash if robotank turret deconstructed by a robot.  Bump to 0.3.2.
If a robot deconstructs a hidden turret, the reference held by the
controller becomes invalid and the mod crashes.  The solution is to
set the 'not-deconstructable' flag on the entity so the planner will
not try to deconstruct it.  (I hit this by accident.  I was trying to
deconstruct some trees near the robotanks and they happened to get
selected too.)

I also set 'not-blueprintable' for good measure, although in my tests,
the hidden turret already refused to go into a blueprint.

I considered adding a defensive validity check for the turret to avoid
the immediate crash if, say, another mod destroys my turret.  But there
isn't really a good thing to do if it happens, so I've left it as is.