control.lua: Fix crash when loading game with RoboTank and populated vehicle.
Bug originally reported here:

  https://mods.factorio.com/mod/RoboTank/discussion/5c9e7484ef0d7a000d0c4326

The problem was due to a saved game that had been used in multiplayer,
where one of the other players was in a vehicle when the game was saved.
On load, events related to players changing vechicles fired so that mods
would be consistent with the depopulated vehicles, but this happened at
a time when I wasn't prepared to handle events yet.

The solution is to initialize my data structures in on_init, which is
what I should have been doing anyway.