unit-tests.sh: Update paths for my current setup.
control.lua: Remove "util" require statement.  Evidently this was
for table.deepcopy, but the code no longer uses it.  This require
statement caused the unit tests to fail because util.lua (in the
Factorio directory) refers to `defines.direction`, but that is nil
in my current unit test setup for some reason.