control.lua: remove_entity_controller: Fix bug if two tanks die in quick succession.
This is an old bug that I just now discovered.  If two tanks die in
quick succession, and are immediately replaced by robots, then the
first to die will remember the second in its nearby-entities list,
but it will be invalid.  The fix is to clear nearby-entities even for
the destroyed entity.  (Not doing so was not really intended as an
optimization, rather there was broken logic that tried to limit the
clearing to one force.  It's wrong either way, but in a slightly less
obvious way.)